pyproject.toml
[project]
name = "rootsage"
version = "0.1.0"
description = ""
authors = [
    {name = "ljberrios"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "joblib (>=1.4.2,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

